all files updated

+++ Demo4/src/main/java/com/demo4/WebController.java
@@ -1,5 +1,7 @@
 package com.demo4;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -22,11 +24,10 @@ public class WebController {
 		return "studentform.jsp";
 	}
 
-	@ResponseBody
 	@RequestMapping("/savestudent")
 	String saveStudent(Student s) {
 		repo.save(s);
-		return "<h2>Student Saved</h2>";
+		return "studentform.jsp";
 	}
 
 	@RequestMapping("/deletestudents")
@@ -34,20 +35,58 @@ public class WebController {
 		repo.deleteAll();
 		return "All Students Deleted Successfully";
 	}
-	
+
 	@RequestMapping("/studentdetails")
-	String getstudentbyid(HttpSession h1,Integer id) {
-		Student student=repo.findById(id).orElse(null);
-		h1.setAttribute("id", student.getId());
-		h1.setAttribute("name", student.getName());
-		h1.setAttribute("email", student.getEmail());
-		h1.setAttribute("phno", student.getPhno());
-		return "studentDetails.jsp";
+	String getstudentbyid(HttpSession h1, Integer id) {
+		Student student = repo.findById(id).orElse(null);
+		if (student != null) {
+			h1.setAttribute("id", student.getId());
+			h1.setAttribute("name", student.getName());
+			h1.setAttribute("email", student.getEmail());
+			h1.setAttribute("phno", student.getPhno());
+			return "studentDetails.jsp";
+		} else {
+			return "errorpage.jsp";
+		}
+
 	}
 	
+	@RequestMapping("/getallstudents")
+	String getallstudents(HttpSession h1,Student s) {
+		List<Student> list=(List<Student>) repo.findAll();
+		h1.setAttribute("students", list);
+		return "allstudents.jsp";
+	}
+
 	@RequestMapping("/deletestudent")
 	String deletestudentbyid(Integer id) {
-		repo.deleteById(id);
-		return "deletestudent.jsp";
+		Student student = repo.findById(id).orElse(null);
+		if (student != null) {
+			repo.deleteById(id);
+			return "deletestudent.jsp";
+		} else {
+			return "errorpage.jsp";
+		}
+	}
+	
+	@ResponseBody
+	@RequestMapping("/getbyemail")
+	List<Student> getByEmail(String email) {
+		List<Student> emails= repo.findByEmail(email);
+		return emails;
+	}
+	
+	@ResponseBody
+	@RequestMapping("/sortbyphno")
+	List<Student> sortByPhno(){
+		List<Student> phnos=repo.findByPhno();
+		return phnos;
+	}
+	
+	@ResponseBody
+	@RequestMapping("/getbyname")
+	List<Student> getByName(String name){
+		List<Student> names=repo.findByName(name);
+		return names;
 	}
 }
